update... merged develop into this branch* and... attempted to resolve most merge conflicits... and... created GratuityNew comp and OrderSummaryNew comp
