+++ src/Components/Gratuity/Gratuity.js
@@ -1,19 +1,5 @@
 import "./_Gratuity.scss";
 import React from "react";
-import { useState } from "react";
-
-function Gratuity({ getGratuitycPercentage }) {
-  const [gratuityTracker, setGratuityTracker] = useState(0);
-
-  function clickHandlerButton(e) {
-    e.preventDefault();
-    setGratuityTracker(e.target.value);
-
-    // i want the Gratuity Component to tell the parent component
-    // the gratuiy percentage
-
-    getGratuitycPercentage(e.target.value);
-  }
 import { Link } from "react-router-dom";
 import { useState } from "react";
 
@@ -27,29 +13,6 @@ function Gratuity({
   const [popUpClass, setPopUpClass] = useState("display-none");
   const [showCustomInput, setShowCustomInput] = useState("");
   return (
-    <div className="gratuitywrapper">
-      {/* <div>{gratuityTracker}</div> */}
-      <form>
-        <div>
-          <button onClick={clickHandlerButton} value="10">
-            10%
-          </button>
-          <button onClick={clickHandlerButton} value="15">
-            15%
-          </button>
-          <button onClick={clickHandlerButton} value="18">
-            18%
-          </button>
-          <button onClick={clickHandlerButton} value="20">
-            20%
-          </button>
-        </div>
-
-        <label>
-          Custom %<input></input>
-        </label>
-      </form>
-    </div>
     <>
       <div className="popup_update">
         <h3>Update Gratuity</h3>
@@ -113,8 +76,6 @@ function Gratuity({
           }}
         ></input>
       </div>
-
-     
     </>
   );
 }
