+++ src/Components/GratuityNew/GratuityNew.js
@@ -1,58 +1,43 @@
-import "./_Gratuity.scss";
+import "./GratuityNew.scss";
 import React from "react";
+import { useState } from "react";
 
-function Gratuity({ customGratuity, changeGratuity }) {
+function Gratuity({ getGratuitycPercentage }) {
+  const [gratuityTracker, setGratuityTracker] = useState(0);
+
+  function clickHandlerButton(e) {
+    e.preventDefault();
+    setGratuityTracker(e.target.value);
+
+    // i want the Gratuity Component to tell the parent component
+    // the gratuiy percentage
+
+    getGratuitycPercentage(e.target.value);
+  }
   return (
-    <>
-      <div className="popup_update">
-        <h3>Update Gratuity</h3>
-      </div>
-      <div className="tipAmount">
-        <p
-          onClick={(event) => {
-            event.preventDefault();
-            changeGratuity(event);
-          }}
-        >
-          10%
-        </p>
-        <p
-          onClick={(event) => {
-            event.preventDefault();
-            changeGratuity(event);
-          }}
-        >
-          15%
-        </p>
-        <div className="tipAmount__default">
-          {/* <p className="tipAmount__default--top">$30.08</p> */}
-          <p
-            onClick={(event) => {
-              event.preventDefault();
-              changeGratuity(event);
-            }}
-            className="tipAmount__default--buttom"
-          >
+    <div className="gratuitywrapper">
+      {/* <div>{gratuityTracker}</div> */}
+      <form>
+        <div>
+          <button onClick={clickHandlerButton} value="10">
+            10%
+          </button>
+          <button onClick={clickHandlerButton} value="15">
+            15%
+          </button>
+          <button onClick={clickHandlerButton} value="18">
             18%
-          </p>
+          </button>
+          <button onClick={clickHandlerButton} value="20">
+            20%
+          </button>
         </div>
-        <p
-          onClick={(event) => {
-            event.preventDefault();
-            changeGratuity(event);
-          }}
-        >
-          20%
-        </p>
-        <p>Custom</p>
-      </div>
-      <div className="buttonBox">
-        <button type="text" className="button">
-          {" "}
-          Confirm 18% ($30.08){" "}
-        </button>
-      </div>
-    </>
+
+        <label>
+          Custom %<input></input>
+        </label>
+      </form>
+    </div>
   );
 }
 export default Gratuity;
