+++ src/Components/OrderSummaryNew/OrderSummaryNew.scss
@@ -1,0 +1,29 @@
+@use "../../styles/partials/fonts" as *;
+@use "../../styles/partials/variables.scss" as *;
+@use "../../styles/partials/mixins.scss" as *;
+
+.tablesummarywrapper{
+    background-color:antiquewhite;
+    width: 100%;
+}
+.hero{
+    margin: 16px;
+    color: $textColorLight;
+    &__price{
+        display: flex;
+        justify-content: space-between;
+    }
+    
+}
+.header{
+    margin: 16px;
+}
+.total{
+    margin: 16px;
+}
+.gratuityvisibilitybutton{
+    background-color: transparent;
+    border: none;
+    color: blue;
+    cursor:pointer;
+}
